==> notebooks/exploratory/0_eda_dados_gerados.ipynb <==
# Análise exploratória dos dados gerados

import os

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sistema_antifraude.gerador_dados import Gerador
from sistema_antifraude import config

dados_gerados = Gerador().gerar_dados()

df = pd.DataFrame(dados_gerados)
df.head()

df.info()

df['timestamp'] = pd.to_datetime(df['timestamp'])

df.sort_values(by='timestamp', inplace = True)

df['is_fraude'].value_counts(normalize = True)

usuarios_por_dia = df.groupby(df['timestamp'].dt.date)['user_id'].count()
ax = sns.barplot(data = usuarios_por_dia)
ax.set_title('Quantidade de acessos por dia')
ax.set_xlabel('Dias')
ax.set_ylabel('Acessos')
ax.set_xticklabels(ax.get_xticklabels(), rotation = 90)
plt.show()

usuarios_ao_longo_do_dia = df[df['timestamp'].dt.date == df.loc[0]['timestamp'].date()]
usuarios_ao_longo_do_dia = usuarios_ao_longo_do_dia.groupby(usuarios_ao_longo_do_dia['timestamp'].dt.hour)['user_id'].count()
ax = sns.barplot(data = usuarios_ao_longo_do_dia)
ax.set_title('Quantidade de acessos ao longo do dia')
ax.set_xlabel('Hora')
ax.set_ylabel('Acessos')
plt.show()

qtd_dispositivos_por_usuario = df.groupby('user_id')['device_id'].nunique().value_counts()
ax = sns.barplot(data = qtd_dispositivos_por_usuario)
ax.set_title('Quantidade de dispositivos por usuário')
ax.set_xlabel('Quantidade de dispositivos')
ax.set_ylabel('Quantidade de usuários')
plt.show()

df.to_csv(os.path.join(config.DATA_PATH, 'raw', 'dados_gerados.csv'), index = False)
